+++ typescript/src/lib/api_container_context/api_container_context.ts
@@ -1,0 +1,36 @@
+
+export class ApiContainerContext {
+    private readonly containerId:  string
+    private readonly ipInsideEnclave: string
+    private readonly portInsideEnclave: number
+    private readonly ipOnHostMachine: string
+    private readonly portOnHostMachine: number
+
+    constructor(containerId: string, ipInsideEnclave: string, portInsideEnclave: number, ipOnHostMachine: string, portOnHostMachine: number) {
+        this.containerId = containerId;
+        this.ipInsideEnclave = ipInsideEnclave;
+        this.portInsideEnclave = portInsideEnclave;
+        this.ipOnHostMachine = ipOnHostMachine;
+        this.portOnHostMachine = portOnHostMachine;
+    }
+
+    public getContainerId(): string {
+        return this.containerId;
+    }
+
+    public getIpInsideEnclave(): string {
+        return this.ipInsideEnclave;
+    }
+
+    public getPortInsideEnclave(): number {
+        return this.portInsideEnclave;
+    }
+
+    public getIpOnHostMachine(): string {
+        return this.ipOnHostMachine;
+    }
+
+    public getPortOnHostMachine(): number {
+        return this.portOnHostMachine;
+    }
+}